finzen-server/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database (Async MongoDB)
motor==3.3.2
pymongo==4.6.0

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration & Validation
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS & HTTP
fastapi-cors==0.0.6
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.1

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1